backend/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.24.0.post1
httpx==0.26.0
python-dotenv==1.0.1
python-multipart==0.0.9
